refactor(models/github): 重构组织信息接口和类型定义
- 重命名 OrganizationNameParamType 为 OrgInfoParamType
- 重命名 OrganizationInfoType 为 OrgInfoResponseType
- 移除不必要的类型定义，简化组织信息结构
- 优化 get_org_info 方法，返回精简的组织信息
