+++ package.json
@@ -45,6 +45,7 @@
     "test": "tsx test/index.ts"
   },
   "dependencies": {
+    "@types/lodash": "^4.17.16",
     "axios": "^1.8.4",
     "color-convert": "^3.1.0",
     "dayjs": "^1.11.13",
@@ -53,6 +54,7 @@
     "https-proxy-agent": "^7.0.6",
     "jsonwebtoken": "^9.0.2",
     "language-colors": "^2.1.55",
+    "lodash": "^4.17.21",
     "socks-proxy-agent": "^8.0.5",
     "uuid": "^11.1.0"
   },
