+++ src/models/platform/github/user.ts
@@ -1,4 +1,6 @@
 /* eslint-disable @typescript-eslint/prefer-nullish-coalescing */
+import _ from 'lodash'
+
 import {
   getContributionData,
   isNotAccessTokeMsg,
@@ -73,7 +75,7 @@ export class User extends Base {
           id: res.data.id,
           login: res.data.login,
           name: res.data.name || null,
-          type: res.data.type,
+          type: _.capitalize(res.data.type.toLowerCase()),
           html_url: res.data.html_url,
           avatar_url: res.data.avatar_url,
           company: res.data.company || null,
@@ -127,7 +129,7 @@ export class User extends Base {
           id: res.data.id,
           login: res.data.login,
           name: res.data.name || null,
-          type: res.data.type,
+          type: _.capitalize(res.data.type.toLowerCase()),
           html_url: res.data.html_url,
           avatar_url: res.data.avatar_url,
           company: res.data.company || null,
@@ -177,7 +179,7 @@ export class User extends Base {
           id: res.data.id,
           login: res.data.login,
           name: res.data.name || null,
-          type: res.data.type,
+          type: _.capitalize(res.data.type.toLowerCase()),
           html_url: res.data.html_url,
           avatar_url: res.data.avatar_url,
           company: res.data.company || null,
